requirements.txt
# PDF to Markdown Converter - Requirements
# Python dependencies for converting PDFs and other documents to Markdown using Datalab.to API

# Core dependencies
datalab-python-sdk>=0.2.2
python-dotenv>=1.2.1
tqdm>=4.67.0
